Switch main to the entry instead of the distribution
This makes the assumption that you'll be compiling vcom to work in all browsers at a later time (like in your app). It currently only works natively in Chrome.

It simplified the pipeline a lot to do this, so i think this is a good tradeoff to make, but you can always reference vcom/dist/vcom.js if you want the cross-browser distribution.